# CUDA: Optimize PAD_REFLECT_1D