in the previous PR  #14659  , [JohannesGaessler](https://github.com/JohannesGaessler) said  https://github.com/ggml-org/llama.cpp/pull/14659#discussion_r2206946843
> This is going to produce correct results but generally speaking you will get much better performance if each thread just works on a single value instead of looping over `ne0`. However, it would also be fine to just merge it as-is and maybe change this later if it ever becomes relevant for end-to-end performance.

so I wrote a version without loops. it benefits the most for smaller tensor sizes, but in general it improves on any size as well.

by the way I added more test cases for PAD_REFLECT_1D.

here is the benchmark summary:
---

## Benchmark Results

| Tensor Shape (ne\_a) | Before (GB/s) | After (GB/s) | Improvement |
| -------------------- | ------------- | ------------ | ----------- |
| \[512, 34, 2, 1]     | 70.85         | 91.82        | **+29.6%**  |
| \[3000, 80, 1, 1]    | 450.38        | 570.28       | **+26.6%**  |
| \[3000, 80, 4, 1]    | 447.39        | 571.26       | **+27.7%**  |
| \[3000, 384, 1, 1]   | 339.07        | 362.65       | **+6.9%**   |
| \[3000, 384, 4, 1]   | 339.30        | 362.59       | **+6.9%**   |

---

## Highlights

* Overall **\~7–30% performance improvement** depending on tensor shape.
* Best gains observed on medium-sized tensors (width = 80).
* Large tensors (width = 384) also show consistent improvements (\~7%).

---

### Overall Result

* **Average bandwidth gain across all tensor shapes:** ≈ **+15–16%**.
* **Best improvement:** \~**+21%** on small/medium tensors (`[512,34,2,1]`, `[3000,80,1,1]`, `[3000,80,4,1]`).
* **Moderate improvement:** \~**+7%** on large tensors (`[3000,384,1,1]`, `[3000,384,4,1]`).

---


# raw benchmark data:

## old kernel
```
.\test-backend-ops.exe perf -o PAD_REFLECT_1D -b CUDA0
Backend 1/2: CUDA0
  Device description: NVIDIA GeForce RTX 3080 Laptop GPU
  Device memory: 16383 MB (15253 MB free)

  PAD_REFLECT_1D(type=f32,ne_a=[512,34,2,1],pad_0=10,pad_1=9):                540672 runs -     1.86 us/run -      138 kB/run -   70.85 GB/s
  PAD_REFLECT_1D(type=f32,ne_a=[3000,80,1,1],pad_0=10,pad_1=9):               253952 runs -     3.98 us/run -     1880 kB/run -  450.38 GB/s
  PAD_REFLECT_1D(type=f32,ne_a=[3000,384,1,1],pad_0=10,pad_1=9):               40887 runs -    25.39 us/run -     9028 kB/run -  339.07 GB/s
  PAD_REFLECT_1D(type=f32,ne_a=[3000,80,4,1],pad_0=10,pad_1=9):               253952 runs -     4.01 us/run -     1880 kB/run -  447.39 GB/s
  PAD_REFLECT_1D(type=f32,ne_a=[3000,384,4,1],pad_0=10,pad_1=9):               40887 runs -    25.38 us/run -     9028 kB/run -  339.30 GB/s
```
## new kernel 
```
.\test-backend-ops.exe perf -o PAD_REFLECT_1D 
Backend 1/2: CUDA0
  Device description: NVIDIA GeForce RTX 3080 Laptop GPU
  Device memory: 16383 MB (15253 MB free)

 PAD_REFLECT_1D(type=f32,ne_a=[512,34,2,1],pad_0=10,pad_1=9):                696320 runs -     1.44 us/run -      138 kB/run -   91.82 GB/s
  PAD_REFLECT_1D(type=f32,ne_a=[3000,80,1,1],pad_0=10,pad_1=9):               319488 runs -     3.15 us/run -     1880 kB/run -  570.28 GB/s
  PAD_REFLECT_1D(type=f32,ne_a=[3000,80,4,1],pad_0=10,pad_1=9):               319488 runs -     3.14 us/run -     1880 kB/run -  571.26 GB/s
  PAD_REFLECT_1D(type=f32,ne_a=[3000,384,1,1],pad_0=10,pad_1=9):               44604 runs -    23.74 us/run -     9028 kB/run -  362.65 GB/s
  PAD_REFLECT_1D(type=f32,ne_a=[3000,384,4,1],pad_0=10,pad_1=9):               44604 runs -    23.75 us/run -     9028 kB/run -  362.59 GB/s
```