@yosh20004 2025-09-14
Thanks for the PR — really nice work!

I was exploring a similar implementation and we both use threadblock layout related to **ne0**, but I think that keeping a loop structure in this kernel may hide memory access latency, so I want to introduce a parameter **UNROLL** to control grid-stride loop.

In short, I want **ne0 / UNROLL** per line, each process **UNROLL** elements in a single iteration, and i think your approach implied **UNROLL=1**.

Here are some benchmark results (Device: 3060M 6G):
| Config      | Shape (ne_a)       | Runs    | Time (us/run) | Size (kB/run) | Bandwidth (GB/s) | Speedup vs PR |
|-------------|-------------------|---------|---------------|---------------|------------------|---------------|
| **UNROLL=4** | [512,34,2,1]      | 491520  | 2.04          | 138           | 64.83            | 0.84x         |
|             | [3000,80,1,1]     | 229376  | 4.45          | 1880          | 403.39           | 1.28x         |
|             | [3000,384,1,1]    | 29736   | 35.02         | 9028          | 245.88           | 1.19x         |
|             | [3000,80,4,1]     | 229376  | 4.47          | 1880          | 401.21           | 1.37x         |
|             | [3000,384,4,1]    | 29736   | 35.07         | 9028          | 245.50           | 1.19x         |
| **UNROLL=1** | [512,34,2,1]      | 565248  | 1.78          | 138           | 74.18            | 0.96x         |
|             | [3000,80,1,1]     | 155648  | 6.67          | 1880          | 269.07           | 0.85x         |
|             | [3000,384,1,1]    | 22302   | 46.04         | 9028          | 187.02           | 0.91x         |
|             | [3000,80,4,1]     | 147456  | 6.85          | 1880          | 261.91           | 0.89x         |
|             | [3000,384,4,1]    | 26019   | 42.07         | 9028          | 204.67           | 0.99x         |
| **Your PR** | [512,34,2,1]      | 589824  | 1.71          | 138           | 77.45            | 1.00x         |
|             | [3000,80,1,1]     | 180224  | 5.69          | 1880          | 315.06           | 1.00x         |
|             | [3000,384,1,1]    | 26019   | 41.79         | 9028          | 206.06           | 1.00x         |
|             | [3000,80,4,1]     | 172032  | 6.12          | 1880          | 293.23           | 1.00x         |
|             | [3000,384,4,1]    | 26019   | 41.55         | 9028          | 207.21           | 1.00x         |

not a big change, but it might improve robustness across workloads.

If you think it makes sense, you can take a look at my draft implementation:
https://github.com/yosh20004/llama.cpp/blob/feat/optimize-pad_reflect_1d/ggml/src/ggml-cuda/pad_reflect_1d.cu