@yosh20004 2025-09-14
> On some GPUs it's faster to pass float pointers and to calculate offsets like `s01 = nb01 / sizeof(float)` in host code.

You are right. If the tensor is small, I guess we need to activate more cuda core to increase the occupancy of SM. Maybe we can have a strategy to set different UNROLL value according to the tensor size, but alse it will make code more complex.

Such as:

```cpp
template<int UNROLL=1>
static __global__ void pad_reflect_1d_kernel_f32(...);

void ggml_cuda_op_pad_reflect_1d(ggml_backend_cuda_context & ctx, ggml_tensor * dst) {
    // same as your code
    if (TENSORSIZE <= LIMIT) {
        pad_reflect_1d_kernel_f32<1><<<grid, block, 0, stream>>>(...); // better for small tensor
    } else {
        pad_reflect_1d_kernel_f32<4><<<grid, block, 0, stream>>>(...); // better for large tensor
    }
}
```

I think the amount of **LIMIT** is related to the GPU architecture, in my device its about 32768 (i0 * i1 * i2 * i3), but I am not sure.